+++ src/utils/logger_setup.py
@@ -1,0 +1,7 @@
+import logging
+
+
+def setup_logger():
+    logging.basicConfig(level=logging.DEBUG)
+    logger = logging.getLogger(__name__)
+    return logger
